+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+* Thrown exceptions implement `Gamez\Duration\DurationException`
+* An operation resulting in an invalid duration will throw a `Gamez\Duration\Exception\InvalidDuration`
+
 ## 1.0.1 - 2019-01-24
 
 * Fixed errors in the documentation